Cleanup: Move some more Modals into common/Modals
