+++ ui/src/components/common/Modals/index.ts
@@ -1,3 +1,7 @@
+export * from './ConfirmModal';
+export * from './ErrorModal';
+export * from './ErrorModalItem';
+export * from './JoreErrorModal';
 export * from './Modal';
 export * from './ModalBody';
 export * from './ModalHeader';
